refactor(#3272): extract extern-declarations pre-pass from index.ts (refs #3182)
Built-in extern-class registration + declare/interface/mixin/namespace/enum/
global collection + node-builtin/jsx-runtime import registration + WASI/DOM
usage guards moved verbatim. index.ts imports back the 9 driver entry points +
re-exports the 4 public entry points. Byte-identity IDENTICAL (39/39), tsc 0.

✓
